Added transcoding, this will make more sense later
Added the transcoders folder for these transcoders.
Made the parroting modes return the full entry instead of just what was previously required.
Made the source response from the server only respond with the body, not the headers
Added an api call to change the transcode.
Added a preview page so I can actually write to that without getting security exceptions.
Trying to display the source of the response does crash FF right now though...